RC41: Temporarily disable flash on boot, clean up logs, fix incorrect module names
